+++ test/Doctrine/EntityManagerFactoryTest.php
@@ -7,6 +7,7 @@ namespace ShlinkioTest\Shlink\Common\Doctrine;
 use Doctrine\DBAL\Driver\PDO\SQLite\Driver as SQLiteDriver;
 use Doctrine\DBAL\Types\Type;
 use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\Events;
 use Doctrine\Persistence\Mapping\Driver\PHPDriver;
 use Laminas\ServiceManager\ServiceManager;
 use PHPUnit\Framework\TestCase;
@@ -14,10 +15,12 @@ use ReflectionObject;
 use Shlinkio\Shlink\Common\Doctrine\EntityManagerFactory;
 use Shlinkio\Shlink\Common\Doctrine\Type\ChronosDateTimeType;
 use ShlinkioTest\Shlink\Common\Repository\CustomRepository;
+use stdClass;
 
 use function array_filter;
 use function array_merge;
 use function array_merge_recursive;
+use function count;
 
 use const ARRAY_FILTER_USE_KEY;
 
@@ -47,10 +50,17 @@ class EntityManagerFactoryTest extends TestCase
      * @test
      * @dataProvider provideConfig
      */
-    public function serviceIsCreated(array $config, int $expectedAutoGenerateProxies, string $expectedDefaultRepo): void
-    {
+    public function serviceIsCreated(
+        array $config,
+        int $expectedAutoGenerateProxies,
+        string $expectedDefaultRepo,
+        int $expectedListeners
+    ): void {
         $sm = new ServiceManager(['services' => [
             'config' => $config,
+            'foo_listener' => new stdClass(),
+            'bar_listener' => new stdClass(),
+            'baz_listener' => new stdClass(),
         ]]);
 
         self::assertFalse(Type::hasType(ChronosDateTimeType::CHRONOS_DATETIME));
@@ -65,6 +75,19 @@ class EntityManagerFactoryTest extends TestCase
         /** @var PHPDriver $metaDriver */
         $metaDriver = $em->getConfiguration()->getMetadataDriverImpl();
         self::assertEquals([__FILE__], $metaDriver->getLocator()->getPaths());
+
+        $events = $em->getEventManager();
+        $ref = new ReflectionObject($events);
+        $prop = $ref->getProperty('_listeners');
+        $prop->setAccessible(true);
+        $listeners = $prop->getValue($events);
+
+        $listenersCount = 0;
+        foreach ($listeners as $list) {
+            $listenersCount += count($list);
+        }
+
+        self::assertEquals($expectedListeners, $listenersCount);
     }
 
     public function provideConfig(): iterable
@@ -84,12 +107,12 @@ class EntityManagerFactoryTest extends TestCase
             ],
         ];
 
-        yield [array_merge($baseConfig, ['debug' => true]), 1, EntityRepository::class];
-        yield [array_merge($baseConfig, ['debug' => '1']), 1, EntityRepository::class];
-        yield [array_merge($baseConfig, ['debug' => 'true']), 1, EntityRepository::class];
-        yield [array_merge($baseConfig, ['debug' => false]), 0, EntityRepository::class];
-        yield [array_merge($baseConfig, ['debug' => null]), 0, EntityRepository::class];
-        yield [array_merge($baseConfig, ['debug' => null]), 0, EntityRepository::class];
+        yield [array_merge($baseConfig, ['debug' => true]), 1, EntityRepository::class, 0];
+        yield [array_merge($baseConfig, ['debug' => '1']), 1, EntityRepository::class, 0];
+        yield [array_merge($baseConfig, ['debug' => 'true']), 1, EntityRepository::class, 0];
+        yield [array_merge($baseConfig, ['debug' => false]), 0, EntityRepository::class, 0];
+        yield [array_merge($baseConfig, ['debug' => null]), 0, EntityRepository::class, 0];
+        yield [array_merge($baseConfig, ['debug' => null]), 0, EntityRepository::class, 0];
         yield [
             array_merge_recursive($baseConfig, [
                 'entity_manager' => [
@@ -98,6 +121,22 @@ class EntityManagerFactoryTest extends TestCase
             ]),
             0,
             CustomRepository::class,
+            0,
+        ];
+        yield [
+            array_merge_recursive($baseConfig, [
+                'entity_manager' => [
+                    'orm' => [
+                        'listeners' => [
+                            Events::postFlush => ['foo_listener', 'bar_listener'],
+                            Events::prePersist => ['foo_listener', 'bar_listener', 'baz_listener'],
+                        ],
+                    ],
+                ],
+            ]),
+            0,
+            EntityRepository::class,
+            5,
         ];
     }
 }